Move accounts routes to separate directory
